requirements.txt
transformers==4.34.0
tensorflow==2.12.0
tensorflow-cpu==2.12.0
numpy<2.0
sentencepiece==0.1.99
python-telegram-bot==20.3